Move dataset transform functions into the dataset
