--- r2c_isg/functions/__init__.py
@@ -1,10 +1,0 @@
-from .trim import trim
-from .sample import sample
-from .sort import sort
-
-
-function_map = {
-    'trim': trim,
-    'sample': sample,
-    'sort': sort
-}
